requirements.txt
google-generativeai
pytest
python-dotenv
ollama
